# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PAW Blockchain API
  version: 1.0.0
  description: |
    # PAW Blockchain REST API Documentation

    Complete API reference for the PAW Blockchain - a lean Layer-1 blockchain with built-in DEX,
    secure compute aggregation, and mobile-ready wallets.

    ## Features
    - **Native DEX**: Atomic swaps and AMM pools
    - **Oracle Module**: Decentralized price feeds
    - **Compute Module**: TEE-protected API aggregation
    - **Standard Cosmos SDK**: Bank, staking, governance, auth endpoints

    ## Base URLs
    - **Mainnet**: `https://api.paw.network`
    - **Testnet**: `https://testnet-api.paw.network`
    - **Local**: `http://localhost:1317`

    ## Authentication
    Most read endpoints are public. Write operations require transaction signing with your private key.

    ## Rate Limits
    - Public endpoints: 100 requests/minute
    - Authenticated endpoints: 1000 requests/minute

    ## Versioning
    Current API version: v1. Breaking changes will increment the major version.

  contact:
    name: PAW Blockchain Team
    url: https://paw.network
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:1317
    description: Local development server
  - url: https://testnet-api.paw.network
    description: Testnet server
  - url: https://api.paw.network
    description: Mainnet production server
...
paths:
...
  /cosmos/bank/v1beta1/send:
    post:
      tags: [Bank]
      summary: Send tokens
      description: Transfer tokens to another address
      operationId: sendTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
      responses:
        '200':
          description: Transaction broadcast successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TxResponse'

  # Staking Endpoints
...
components:
  schemas:
...
    Coin:
      type: object
      properties:
        denom:
          type: string
          example: "uapaw"
        amount:
          type: string
          example: "1000000"
...
    TxResponse:
      type: object
      properties:
        height:
          type: integer
          format: int64
        txhash:
          type: string
        codespace:
          type: string
        code:
          type: integer
        data:
          type: string
        raw_log:
          type: string
        logs:
          type: array
          items:
            type: object
        info:
          type: string
        gas_wanted:
          type: integer
          format: int64
        gas_used:
          type: integer
          format: int64
        tx:
          type: object
        timestamp:
          type: string
...
    SendRequest:
      type: object
      required: [from_address, to_address, amount]
      properties:
        from_address:
          type: string
        to_address:
          type: string
        amount:
          type: array
          items:
            $ref: '#/components/schemas/Coin'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authenticated endpoints (optional for public data)
security:
  - ApiKeyAuth: []